Implement player speed modification through power-ups
A PowerUp system has been implemented that allows applying power-ups to players that temporarily modify their attributes. Currently, only a speed modification power-up is available, but the system is easily expandable for other types of power-ups. Related changes across the codebase include adjustments to the Player class and its methods, initialization and application of the SpeedPowerUp in the main loop, and addition of power-up texture assets.