Adjust player width and clean up code
The player's width in the Player class was decreased from 15.0f to 1.0f for a slimmer design. Plus a minor typo fix in the main.cpp file. In the Ball class, an additional randomness was added to the angle calculated after the ball intersects with the player for more unpredictable play.